Fix some edge cases with the toolbar lozenge accessory
Apparently it is not a good idea to use the 'isHidden' property;
although I can't detect any problems in my own apps, when I tried to
retrofit TextEdit.app to use AquaWindow, this caused the window title to
get laid out behind the traffic buttons. I am not certain what the cause
of this is, but suffice it to say that the new code fixes the problem at
the cost of some complexity.